requirements.txt
# =============================================================================
# Worker Connect API - Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Core Django
# -----------------------------------------------------------------------------
Django==4.2.17
djangorestframework==3.16.1
django-cors-headers==4.9.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# -----------------------------------------------------------------------------
# API & Documentation
# -----------------------------------------------------------------------------
drf-yasg==1.21.8
PyYAML==6.0.2

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
django-ratelimit==4.1.0
python-magic==0.4.27

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
python-decouple==3.8

# -----------------------------------------------------------------------------
# Frontend Integration
# -----------------------------------------------------------------------------
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0

# -----------------------------------------------------------------------------
# File Processing
# -----------------------------------------------------------------------------
Pillow==11.0.0
reportlab==4.0.7
weasyprint==62.3

# -----------------------------------------------------------------------------
# System Monitoring
# -----------------------------------------------------------------------------
psutil==5.9.8

# -----------------------------------------------------------------------------
# Task Queue (Celery)
# -----------------------------------------------------------------------------
celery==5.4.0
redis==5.0.0

# -----------------------------------------------------------------------------
# WebSocket Support (Django Channels)
# -----------------------------------------------------------------------------
channels==4.1.0
channels-redis==4.2.0

# -----------------------------------------------------------------------------
# Production Server
# -----------------------------------------------------------------------------
gunicorn==22.0.0
uvicorn[standard]==0.30.0

# -----------------------------------------------------------------------------
# Compression
# -----------------------------------------------------------------------------
brotli==1.1.0

# -----------------------------------------------------------------------------
# Development & Testing (optional)
# -----------------------------------------------------------------------------
# django-debug-toolbar==4.3.0
# pytest==8.2.0
# pytest-django==4.8.0
# pytest-cov==5.0.0
# factory-boy==3.3.0
# faker==25.0.0

